backend/requirements.txt
authlib>=1.3.0
flask>=3.0.0
flask-cors>=5.0.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
pypdf>=5.0.0
openai>=1.55.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
